1.增加批量出库接口。

+++ business-center/wms-center/wms-service-api/wms-business-delivery-api/src/main/java/com/szmsd/delivery/dto/ContainerDetailDto.java
@@ -18,5 +18,5 @@ public class ContainerDetailDto implements Serializable {
     private String sku;
 
     @ApiModelProperty(value = "qty")
-    private Integer qty;
+    private Long qty;
 }
